requirements.txt
asgiref==3.2.10
camelcase==0.2
Django==3.1.2
djangorestframework==3.12.2
gunicorn==20.0.4
python-decouple==3.3
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
